requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
charset-normalizer==3.1.0
discord==2.3.1
discord.py==2.3.1
frozenlist==1.3.3
fuzzywuzzy==0.18.0
idna==3.4
Levenshtein==0.21.1
multidict==6.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-Levenshtein==0.21.1
rapidfuzz==3.1.2
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlalchemy-cockroachdb==2.0.1
typing_extensions==4.7.1
yarl==1.9.2